Introduce experimental cache optimization
This change makes the custom control to cache the render points necessary for drawing the control itself. When OnPaint() is called and no property that is a factor for rendering has changed (borderRadius for example), OnPaint() will not redraw the control entirely, instead, it will simply reuse the previously calculated data points and recolor the control. This change should (in theory) improve rendering performance for this control.